Created clean structure with diffrent files
